Fixing Broken tests
preventing tests from failing by adding a the missing browser APIs in the test runs.

Since the ResizeObserver mock code in (`src/setupTests.js` line 22-40) only works on the tests in browser-specific like conditions; tests containing the window object.
Therefore the node-like environment that is created in the `EndToEnd.test.js` file, is not able to run this code- hence failed test.